- Make all JSON test files follow the same format - Add test documentation This is in preparation for moving the test files to the pdf-association github repo
