+++ test/basic.json
@@ -1,100 +1,98 @@
-[{
-   "comment" : "dot operator in object",
-    "given":
-        {"foo": {"bar": {"baz": "correct"}}},
-     "cases": [
-         {
-            "expression": "foo",
-            "result": {"bar": {"baz": "correct"}}
-         },
-         {
-            "expression": "foo.bar",
-            "result": {"baz": "correct"}
-         },
-         {
-            "expression": "foo.bar.baz",
-            "result": "correct"
-         },
-         {
-            "expression": "foo\n.\nbar\n.baz",
-            "result": "correct"
-         },
-         {
-            "expression": "foo.bar.baz.bad",
-            "result": null
-         },
-         {
-            "expression": "foo.bar.bad",
-            "result": null
-         },
-         {
-            "expression": "foo.bad",
-            "result": null
-         },
-         {
-            "expression": "bad",
-            "result": null
-         },
-         {
-            "expression": "bad.morebad.morebad",
-            "result": null
-         }
-     ]
-},
-{
-   "comment" : "dot operator in object (2)",
-    "given":
-        {"foo": {"bar": ["one", "two", "three"]}},
+[
+  {
+    "comment": "dot operator in object",
+    "given": { "foo": { "bar": { "baz": "correct" } } },
     "cases": [
-         {
-            "expression": "foo",
-            "result": {"bar": ["one", "two", "three"]}
-         },
-         {
-            "expression": "foo.bar",
-            "result": ["one", "two", "three"]
-         }
+      {
+        "expression": "foo",
+        "result": { "bar": { "baz": "correct" } }
+      },
+      {
+        "expression": "foo.bar",
+        "result": { "baz": "correct" }
+      },
+      {
+        "expression": "foo.bar.baz",
+        "result": "correct"
+      },
+      {
+        "expression": "foo\n.\nbar\n.baz",
+        "result": "correct"
+      },
+      {
+        "expression": "foo.bar.baz.bad",
+        "result": null
+      },
+      {
+        "expression": "foo.bar.bad",
+        "result": null
+      },
+      {
+        "expression": "foo.bad",
+        "result": null
+      },
+      {
+        "expression": "bad",
+        "result": null
+      },
+      {
+        "expression": "bad.morebad.morebad",
+        "result": null
+      }
     ]
-},
-{
-   "comment" : "dot operator in array",
+  },
+  {
+    "comment": "dot operator in object (2)",
+    "given": { "foo": { "bar": ["one", "two", "three"] } },
+    "cases": [
+      {
+        "expression": "foo",
+        "result": { "bar": ["one", "two", "three"] }
+      },
+      {
+        "expression": "foo.bar",
+        "result": ["one", "two", "three"]
+      }
+    ]
+  },
+  {
+    "comment": "dot operator in array",
     "given": ["one", "two", "three"],
     "cases": [
-        {
-            "expression": "one",
-            "result": null
-        },
-        {
-            "expression": "two",
-            "result": null
-        },
-        {
-            "expression": "three",
-            "result": null
-        },
-        {
-            "expression": "one.two",
-            "result": null
-        }
+      {
+        "expression": "one",
+        "result": null
+      },
+      {
+        "expression": "two",
+        "result": null
+      },
+      {
+        "expression": "three",
+        "result": null
+      },
+      {
+        "expression": "one.two",
+        "result": null
+      }
     ]
-},
-{
-   "comment" : "property access with numeric names",
-    "given":
-        {"foo": {"1": ["one", "two", "three"], "-1": "bar"}},
+  },
+  {
+    "comment": "property access with numeric names",
+    "given": { "foo": { "1": ["one", "two", "three"], "-1": "bar" } },
     "cases": [
-         {
-            "expression": "foo.'1'",
-            "result": ["one", "two", "three"]
-         },
-         {
-            "expression": "foo.'1'[0]",
-            "result": "one"
-         },
-         {
-            "expression": "foo.'-1'",
-            "result": "bar"
-         }
+      {
+        "expression": "foo.'1'",
+        "result": ["one", "two", "three"]
+      },
+      {
+        "expression": "foo.'1'[0]",
+        "result": "one"
+      },
+      {
+        "expression": "foo.'-1'",
+        "result": "bar"
+      }
     ]
-}
+  }
 ]
