+++ test/boolean.json
@@ -77,7 +77,7 @@
       "Number": 5,
       "EmptyList": [],
       "EmptyObject": {},
-      "Obj": {"oneEntry": "one"},
+      "Obj": { "oneEntry": "one" },
       "Array": ["oneEntry"],
       "Zero": 0,
       "nested": {
