+++ test/jsonformula.spec.js
@@ -42,7 +42,14 @@ function toTestFmt(nm) {
     tst.cases.forEach(c => {
       const s = [c.comment || comment, c.expression].filter(x => typeof x === 'string' && x).join(' -> ');
       tests.push([s, {
-        given, expression: c.expression, result: c.result, error: c.error,
+        given,
+        expression: c.expression,
+        result: c.result,
+        error: c.error,
+        data: c.data,
+        fieldsOnly: c.fieldsOnly,
+        language: c.language,
+        precedence: c.precedence,
       }]);
     });
   });
@@ -51,47 +58,132 @@ function toTestFmt(nm) {
 
 const jsonFormula = new JsonFormula(functions, stringToNumber);
 
+function getData(tst) {
+  let data = tst.given;
+  if (tst.data) {
+    if (typeof tst.data === 'string') {
+      data = jsonFormula.search(tst.data, tst.given, {}, 'en-US');
+    } else {
+      data = tst.data;
+    }
+  }
+  return data;
+}
+
+const globals = {
+  $days: [
+    'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday',
+  ],
+  $: 42,
+  $$: 43,
+};
+
+/*
+Register the 'summarize' function.  Given:
+{"array": ["a", "b", "c", "d", "a", "b"]}
+The expression:
+summarize(array)
+returns:
+{
+  "a": 2,
+  "b": 2,
+  "c": 1,
+  "d": 1
+}
+*/
+jsonFormula.search(
+  `register("summarize",
+    &reduce(
+      @,
+      &merge(accumulated, fromEntries([[current, 1 + value(accumulated, current)]])),
+      fromEntries(map(@, &[@, 0]))
+    )
+  )`,
+  {},
+);
+
+/*
+Register the 'localDate' function.  Given:
+"yyyy-mm-dd" return a date value.
+*/
+jsonFormula.search(
+  `register(
+    "localDate",
+    &split(@, "-") | datetime(@[0], @[1], @[2]))`,
+  {},
+);
+
+/*
+Register the 'product' function to multiply two parameters.
+product([4,5]) // 20
+*/
+jsonFormula.search(
+  'register("product", &@[0] * [1])',
+  {},
+);
+
 function executeTest(desc, tst) {
+  if (tst.fieldsOnly) return;
   const grammarResult = testGrammar(tst.expression);
   if (grammarResult === 'SyntaxError') expect(tst.error).not.toBeUndefined();
   if (tst.error === 'SyntaxError') expect(grammarResult).toBe('SyntaxError');
 
   let result;
+  const data = getData(tst);
+  globals.$form = data;
+  const language = tst.language || 'en-US';
+
   try {
-    result = jsonFormula.search(tst.expression, tst.given);
+    result = jsonFormula.search(tst.expression, data, globals, language);
   } catch (e) {
     expect(tst.error).not.toBeUndefined();
     expect(e.name).toBe(tst.error);
     return;
   }
-  expect(result).toEqual(tst.result === undefined ? tst.error : tst.result);
+  if (typeof result === 'number') {
+    expect(result).toBeCloseTo(tst.result, 5);
+  } else {
+    expect(result).toEqual(tst.result);
+  }
 }
 
 function executeTestWithFields(desc, tst) {
-  const root = createForm(tst.given);
+  const data = getData(tst);
+  const root = createForm(data);
   let result;
+  globals.$form = data;
+  const language = tst.language || 'en-US';
+
   try {
     const jsonResult = jsonFormula.search(
       tst.expression,
       root,
-      { $form: root, $: {} },
+      globals,
+      language,
     );
+    // stringify/parse so that the comparison doesn't get confused by field objects
     result = JSON.parse(JSON.stringify(jsonResult));
   } catch (e) {
     expect(tst.error).not.toBeUndefined();
+    expect(e.name).toBe(tst.error);
     return;
   }
-  expect(result).toEqual(tst.result === undefined ? tst.error : tst.result);
+  // expect(result).toEqual(tst.result === undefined ? tst.error : tst.result);
+  if (typeof result === 'number') {
+    expect(result).toBeCloseTo(tst.result, 5);
+  } else {
+    expect(result).toEqual(tst.result);
+  }
 }
 
 const suites = [
   'basic', 'boolean',
-  'current', 'escape',
+  'current', 'docSamples', 'escape',
   'filters', 'functions',
   'identifiers', 'indices',
   'literal', 'multiselect',
-  'pipe', 'slice',
-  'syntax', 'unicode',
+  'pipe', 'slice', 'specSamples',
+  'syntax', 'tests', 'unicode',
   'wildcard',
 ];
 
@@ -108,5 +200,23 @@ suites.forEach(suite => {
     if (tests.length) {
       test.each(tests)('%s', executeTestWithFields);
     }
+  });
+});
+
+describe('test precedence', () => {
+  const tests = toTestFmt('precedence');
+  test.each(tests)('%s', (_desc, tst) => {
+  // check that the precedence computed by antlr matches our implementation
+    const grammarResult = testGrammar(tst.expression);
+    const tree = grammarResult.toStringTree()
+    // clean up extra brackets to make the output easier to read
+      .replace(/\[[0-9 ]*\]/g, '')
+      .replace(/ /g, '')
+      .replace(/\(\(([^(]+)\)\)/g, '$1')
+      .replace(/\(([`0-9]+)\)/g, '$1')
+      .replace(/\(\((true\(\)|false\(\))\)\)/g, '$1')
+      .replace(/^\(\(/, '')
+      .replace(/\)<EOF>\)/, '');
+    expect(tree).toBe(tst.precedence);
   });
 });
