--- test/sampleData.json
@@ -1,53 +1,0 @@
-{
-  "samples": {
-    "purchase-order": {
-      "address": {
-        "street": "123 Oak Street",
-        "city": "San Jose",
-        "state": "California",
-        "country": "USA"
-      },
-      "items": [
-        {
-          "desc": "pens",
-          "quantity": 2,
-          "price": 3.23,
-          "subtotal": 6.46
-        },
-        {
-          "desc": "pencils",
-          "quantity": 3,
-          "price": 1.34,
-          "subtotal": 4.02
-        }
-      ]
-    },
-    "datetime": {
-      "epoch": {
-        "date": 1,
-        "month": 1,
-        "year": 1970,
-        "value": 0
-      },
-      "noon": {
-        "hour": 12,
-        "minute": 0,
-        "second": 0
-      }
-    },
-    "casefold": {
-      "test1": {"str1": "Σίσυφος", "str2": "ΣΊΣΥΦΟΣ"},
-      "test2": {"str1": "Straße", "str2": "Strasse"},
-      "test3": {"str1": "I", "str2": "ı"},
-      "test4": {"str1": "İ", "str2": "i"}
-    },
-    "datedif": {
-      "2022-06-20": "2022-06-20",
-      "2022-06-21": "2022-06-21",
-      "2022-07-21": "2022-07-21",
-      "2023-06-20": "2023-06-20",
-      "2023-06-21": "2023-06-21",
-      "2030-08-01": "2030-08-01"
-    }
-  }
-}
